update(deploy): added gh-pages to deploy on GitHub, also added support for decimal point (float)

+++ src/helpers/brain.js
@@ -4,34 +4,34 @@ const brain = (input = '') => {
     for (let i = 0; i < value.length; i++) {
         if (value[i] === "+") {
             if (tempAns) {
-                tempAns += value[i + 1] ? parseInt(value[i + 1], 10) : 0;
+                tempAns += value[i + 1] ? parseFloat(value[i + 1], 10) : 0;
             } else {
-                tempAns += (parseInt(value[i - 1], 10) + parseInt(value[i + 1], 10));
+                tempAns += (parseFloat(value[i - 1], 10) + parseFloat(value[i + 1], 10));
             }
         }
         if (value[i] === "-") {
             if (tempAns) {
-                tempAns -= value[i + 1] ? parseInt(value[i + 1], 10) : 0;
+                tempAns -= value[i + 1] ? parseFloat(value[i + 1], 10) : 0;
             } else {
-                tempAns += ((value[i - 1] ? parseInt(value[i - 1], 10) : 0) - (value[i + 1] ? parseInt(value[i + 1], 10) : 0));
+                tempAns += ((value[i - 1] ? parseFloat(value[i - 1], 10) : 0) - (value[i + 1] ? parseFloat(value[i + 1], 10) : 0));
             }
         }
         if (value[i] === "*") {
             if (tempAns) {
-                tempAns *= value[i + 1] ? parseInt(value[i + 1], 10) : 1;
+                tempAns *= value[i + 1] ? parseFloat(value[i + 1], 10) : 1;
             } else {
-                tempAns += ((value[i - 1] ? parseInt(value[i - 1], 10) : 1) * (value[i + 1] ? parseInt(value[i + 1], 10) : 1));
+                tempAns += ((value[i - 1] ? parseFloat(value[i - 1], 10) : 1) * (value[i + 1] ? parseFloat(value[i + 1], 10) : 1));
             }
         }
         if (value[i] === "÷" || value[i] === "/") {
             if (tempAns) {
-                tempAns /= value[i + 1] ? parseInt(value[i + 1], 10) : 1;
+                tempAns /= value[i + 1] ? parseFloat(value[i + 1], 10) : 1;
             } else {
-                if (parseInt(value[i + 1]) === 0) {
+                if (parseFloat(value[i + 1]) === 0) {
                     tempAns = 0 // return zero(0) instead of infinity when divided by zero
                     continue
                 }
-                tempAns += ((value[i - 1] ? parseInt(value[i - 1], 10) : 1) / (value[i + 1] ? parseInt(value[i + 1], 10) : 1));
+                tempAns += ((value[i - 1] ? parseFloat(value[i - 1], 10) : 1) / (value[i + 1] ? parseFloat(value[i + 1], 10) : 1));
             }
         }
